[#51] fix: SwipeView 진입 즉시 데이터 로드하도록 수정
