+++ lib/features/swipe_view/presentation/screen/swipe_view.dart
@@ -5,6 +5,7 @@ import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:wehavit/common/utils/emoji_assets.dart';
 import 'package:wehavit/features/swipe_view/presentation/model/swipe_view_model.dart';
 import 'package:wehavit/features/swipe_view/presentation/provider/swipe_view_model_provider.dart';
+import 'package:wehavit/features/swipe_view/presentation/screen/widget/emoji_sheet_widget.dart';
 import 'package:wehavit/features/swipe_view/presentation/screen/widget/reaction_camera_widget.dart';
 import 'package:wehavit/features/swipe_view/presentation/screen/widget/swipe_view_cell.dart';
 
@@ -22,12 +23,11 @@ class SwipeViewState extends ConsumerState<SwipeView> {
   late final SwipeViewModelProvider _swipeViewModelProvider;
 
   bool _initOccurred = false;
+  bool _draw = false;
 
   @override
   void initState() {
     super.initState();
-
-    setState(() {});
   }
 
   @override
@@ -38,7 +38,11 @@ class SwipeViewState extends ConsumerState<SwipeView> {
       _swipeViewModel = ref.watch(swipeViewModelProvider);
       _swipeViewModelProvider = ref.read(swipeViewModelProvider.notifier);
       _swipeViewModelProvider.initializeCamera();
-      ref.read(swipeViewModelProvider.notifier).getTodayConfirmPostModelList();
+      await ref
+          .read(swipeViewModelProvider.notifier)
+          .getTodayConfirmPostModelList();
+
+      setState(() {});
       _initOccurred = true;
     }
   }
@@ -47,20 +51,12 @@ class SwipeViewState extends ConsumerState<SwipeView> {
   Widget build(BuildContext context) {
     return Scaffold(
       // appBar: AppBar(),
-      resizeToAvoidBottomInset: false,
-      floatingActionButtonLocation: FloatingActionButtonLocation.miniCenterTop,
-      floatingActionButton: FloatingActionButton(
-        mini: true,
-        onPressed: () async {
-          await ref
-              .read(swipeViewModelProvider.notifier)
-              .getTodayConfirmPostModelList();
-          setState(() {});
-        },
-      ),
+      // resizeToAvoidBottomInset: false,
+
       body: Stack(
         children: [
           SafeArea(
+            // child: Container(color: Colors.blue),
             child: _swipeViewModel.confirmPostModelList.fold(
               (failure) => Container(
                 color: Colors.cyan,
@@ -317,102 +313,6 @@ class SwipeViewState extends ConsumerState<SwipeView> {
           }.entries,
         );
       },
-    );
-  }
-}
-
-class ShootEmojiWidget extends StatefulWidget {
-  ShootEmojiWidget({
-    super.key,
-    required this.disposeWidgetFromParent,
-    required this.currentPos,
-    required this.targetPos,
-    required this.emojiIndex,
-  });
-  int emojiIndex;
-  Function disposeWidgetFromParent;
-  Point currentPos;
-  Point targetPos;
-
-  @override
-  State<ShootEmojiWidget> createState() => _ShootEmojiWidgetState();
-}
-
-class _ShootEmojiWidgetState extends State<ShootEmojiWidget>
-    with TickerProviderStateMixin {
-  late AnimationController _controller;
-  late Animation _yAnimation, _xAnimation;
-  late Function disposeWidgetFromParent;
-  late double targetXPos;
-  final double emojiSize = 50;
-
-  @override
-  void initState() {
-    super.initState();
-
-    disposeWidgetFromParent = widget.disposeWidgetFromParent;
-
-    _controller = AnimationController(
-      vsync: this,
-      duration: const Duration(milliseconds: 1500),
-    );
-
-    _yAnimation = Tween(begin: 0.0, end: 1.0).animate(
-      CurvedAnimation(
-        parent: _controller,
-        curve: Curves.easeOut,
-      ),
-    );
-    _xAnimation = Tween(begin: 0.0, end: 1.0).animate(
-      CurvedAnimation(
-        parent: _controller,
-        curve: Curves.easeInOut,
-      ),
-    );
-
-    _controller.forward(from: 0.0);
-    _controller.addListener(
-      () {
-        if (mounted) {
-          setState(() {});
-        }
-      },
-    );
-    _controller.addStatusListener(
-      (status) {
-        if (status == AnimationStatus.completed) {
-          disposeWidgetFromParent(widget.key);
-          // dispose();
-          // super.dispose();
-        }
-      },
-    );
-  }
-
-  @override
-  void dispose() {
-    _controller.dispose();
-    super.dispose();
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    return Positioned(
-      bottom: (widget.targetPos.y * 1.0 * _yAnimation.value) +
-          (400) * (1 - _yAnimation.value),
-      left: widget.targetPos.x.toDouble() * _xAnimation.value +
-          (widget.currentPos.x.toDouble()) * (1 - _xAnimation.value) -
-          emojiSize / 2,
-      child: Opacity(
-        opacity: _controller.status == AnimationStatus.completed ? 0.0 : 1.0,
-        child: Image(
-          width: emojiSize,
-          height: emojiSize,
-          image: AssetImage(
-            Emojis.emojiList[widget.emojiIndex],
-          ),
-        ),
-      ),
     );
   }
 }
